+++ transport/tcp/client.go
@@ -2,10 +2,12 @@ package tcp
 
 import (
 	"context"
+	"fmt"
 	"net"
 	"time"
 
 	"github.com/matrixxsoftware/go-mdd/mdd"
+	"github.com/matrixxsoftware/go-mdd/mdd/field"
 	"github.com/matrixxsoftware/go-mdd/transport"
 )
 
@@ -86,4 +88,35 @@ func (c *ClientTransport) send(request []byte) ([]byte, error) {
 		}
 		return response, nil
 	}
+}
+
+// need this in future when we start sending messages asynchronously
+func extractHopId(containers *mdd.Containers) (uint32, error) {
+	// Get MtxMsg container (key 93)
+	mtxMsg := containers.GetContainer(93)
+	if mtxMsg == nil {
+		return 0, fmt.Errorf("container MtxMsg is missing")
+	}
+
+	// Assume no changes to the position of hopId field
+	f := mtxMsg.GetField(14)
+
+	if f.Data == nil {
+		return 0, fmt.Errorf("hopId field is missing")
+	}
+
+	// Copy the field data to a new field
+	hopIdField := mdd.Field{
+		Data:  f.Data,
+		Type:  field.UInt32,
+		Codec: f.Codec,
+	}
+
+	// Get the value of the field
+	hopId, err := hopIdField.GetValue()
+	if err != nil {
+		return 0, err
+	}
+
+	return hopId.(uint32), nil
 }
